refactor: move helper methods
